fixup! refactor: plugin entries management can be simpler
