+++ packages/plugin-infra/src/types.ts
@@ -565,7 +565,6 @@ export namespace Plugin.SNSAdaptor {
             disabled: boolean
             tooltipHint?: string
             onClick?: (walletConnectedCallback?: () => void) => void
-            popperBoundary?: HTMLElement | null
         }) => JSX.Element | null
         /**
          * Used to order the applications on the board
